requirements.txt
keyboard
mouse
attrs
pyyaml

# these are for tests, but it's just simpler to keep them in here
pytest
